Update README.md
First, I created a header titled "Avengers: Infinity War" which is the title of my favorite movie. Second, I put a link of a photo that shows the poster of my favorite movie. Third, I used a blockquote to show the entire summary of the movie and used an unordered list element to separate each part in different paragraphs.